Combine evaluate metrics and profile metrics into a single benchmark tool
